# internal/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Dockmesh API
  summary: Container management platform — HTTP API
  description: |
    REST API for Dockmesh. All endpoints are versioned under `/api/v1`.

    ## Authentication

    Two bearer-token flavours are accepted on the `Authorization` header:

    - **User JWTs** — obtained via `POST /auth/login`, short-lived (15 min),
      refreshable via `POST /auth/refresh`. Use for interactive UI / CLI sessions.
    - **API tokens** — prefix `dmt_`, long-lived, created under
      Settings → API Tokens. Use for CI, scripts, external integrations.

    Both share the same permission model — a token's role dictates what it can do.

    ## Versioning

    Version `v1` is stable in the sense that we don't remove fields or change
    response shapes without warning. New fields may be added at any time, so
    clients should ignore unknown fields gracefully.

    ## Errors

    Non-2xx responses use a uniform shape: `{"error": "<human-readable message>"}`.
    HTTP status carries the category (400/401/403/404/409/500).

    ## Maintenance

    This spec is hand-maintained. `TestOpenAPIDriftAgainstRoutes` compares the
    registered chi routes against the operations declared here and fails CI on
    mismatch — so every PR that touches a route must also touch this file.
  version: "1.0.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
  contact:
    name: Dockmesh
    url: https://dockmesh.dev
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/refresh:
    post:
      operationId: auth.refresh
      tags: [auth]
      summary: Rotate an access token
      description: |
        Exchanges a refresh token for a fresh access+refresh pair. Refresh
        tokens are rotated on every use. Reusing a previously-rotated token
        revokes the whole session family.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New token pair.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LoginResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401":
          description: Invalid, expired, or reused refresh token.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /users:
    get:
      operationId: users.list
      tags: [users]
      summary: List all local users
      responses:
        "200":
          description: Users.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: users.create
      tags: [users]
      summary: Create a new local user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string, minLength: 1 }
                email: { type: string, format: email }
                password: { type: string, format: password, minLength: 8 }
                role:
                  type: string
                  description: Role name — `admin`, `operator`, `viewer`, or a custom role.
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Username already taken.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /proxy/disable:
    post:
      operationId: proxy.disable
      tags: [proxy]
      summary: Stop the embedded Caddy container
      responses:
        "204": { description: Disabled. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /auth/oauth2/{slug}/callback:
    get:
      operationId: oauth2.callback
      tags: [auth]
      summary: Exchange code for token + fetch userinfo + mint session
      security: []
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "302": { description: Redirect to SPA. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "JWT or dmt_<40-char-base64url>"
      description: |
        Accepts either a user session JWT (from `POST /auth/login`) or an API
        token (prefix `dmt_`, from Settings → API Tokens). The route layer
        checks role permissions after authentication.

  # Schemas used across multiple groups live here. Group-specific schemas
  # live inline under the operation to keep the spec readable when new
  # groups get added.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid credentials
...
    User:
      type: object
      required: [id, username, role, mfa_enabled]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string, format: email }
        role: { type: string, example: operator }
        scope_tags:
          type: array
          items: { type: string }
          description: Host tags this user is scoped to. Empty / omitted = access to all hosts.
        mfa_enabled: { type: boolean }
...
    LoginResult:
      type: object
      description: |
        Either a fresh session (access+refresh+user) or an MFA challenge
        (`mfa_required: true` + `mfa_token`) that clients must complete via
        `POST /auth/mfa` before getting a session.
      properties:
        access_token:
          type: string
          description: "15-minute-TTL JWT. Send as `Authorization: Bearer <token>`."
        refresh_token:
          type: string
          description: "Opaque token for `/auth/refresh`. Rotates on every use."
        user:
          $ref: "#/components/schemas/User"
        mfa_required:
          type: boolean
          description: "True when the user has MFA enabled — use `mfa_token` with `/auth/mfa`."
        mfa_token:
          type: string
          description: "Short-lived challenge token for the `/auth/mfa` completion."
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but the token's role lacks the required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BadRequest:
      description: Malformed request body or invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...